Removed the need to define GMMODULE Reorganized some shit Made CUtlLuaVector work even if you define NO_SDK (Will use std::vector)

+++ trunk/gm_tmysql4_boost/gm_tmysql4/database.cpp
@@ -177,7 +177,7 @@ void Database::DoExecute( Query* query )
 
 		if ( query->GetFlags() & QUERY_FLAG_LASTID )
 		{
-			query->SetLastID( mysql_insert_id( pMYSQL ) );
+			query->SetLastID( (int) mysql_insert_id( pMYSQL ) );
 		}
 	}
